Do some refactoring with userrelation, add some more methods for instances.

+++ Cardboard.Net.Rest/API/Common/Ping.cs
@@ -5,5 +5,5 @@ namespace Cardboard.Net.Rest.API;
 internal class Ping
 {
     [JsonProperty("pong")]
-    public int Pong { get; set; }
+    public ulong Pong { get; set; }
 }
